requirements.txt
requests~=2.26.0
websockets~=11.0.3
faker~=18.13.0
pytest~=6.2.5
pyyaml~=6.0.1
jsonpath~=0.82.2
jinja2~=3.1.3
allure-python-commons~=2.9.45